transcode_data.h: missing cast
* transcode_data.h (o4): add missing cast to get rid of implicit
  signed extension and suppress shift-overflow warnings.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@51510 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ transcode_data.h
@@ -56,7 +56,7 @@ RUBY_SYMBOL_EXPORT_BEGIN
 			       0xffffffffU))
 #define o4(b0,b1,b2,b3)	(PType(((((unsigned char)(b1))<<8)|\
 				(((unsigned char)(b2))<<16)|\
-				(((unsigned char)(b3))<<24)|\
+				(((unsigned int)(unsigned char)(b3))<<24)|\
 				((((unsigned char)(b0))&0x07)<<5)|\
 				FOURbt)&\
 			       0xffffffffU))
